refactor(risk): fold presidio_client.go back into presidio.go
The split made sense while RetryingPIIScanner existed as a separate
PIIScanner impl on top of PresidioClient. Now that the retry orchestrator
lives on PresidioClient itself, presidio.go was just a 30-line interface
file beside a 570-line implementation. One file is easier to navigate.

The PresidioClient definition sits directly under the PIIScanner
interface so the contract and its sole production implementation are
right next to each other; StubPIIScanner stays at the bottom.
